Added and styled github contributors #583
