+++ packages/app/app/components/HelpModal/Contributors/styles.scss
@@ -1,0 +1,9 @@
+.contributors {
+    &_card_container {
+        padding: 1rem;
+    }
+    .meta_text {
+        font-size: .85em !important;
+        word-break: break-all;
+    }
+}
